adiciona endpoints para buscar username e email
