--- apps/autenticacao/models/user.py
@@ -1,7 +1,0 @@
-from django.contrib.auth.models import User
-
-from apps.main.models import SoftDeletionModel
-
-
-class Usuario(User, SoftDeletionModel):
-    pass
